requirements.txt
alembic==1.13.1
asyncpg==0.29.0
bcrypt==4.1.2
email-validator==2.1.1
fastapi==0.110.0
httpx==0.27.0
orjson==3.9.15
passlib==1.7.4
pydantic-settings==2.2.1
PyJWT==2.8.0
SQLAlchemy==2.0.28
uvicorn==0.29.0
